added basic functionality for adding, viewing, editing, and deleting workouts

+++ db/migrate/20120402021451_add_reps_and_weight_to_exercise_set.rb
@@ -1,7 +1,7 @@
 class AddRepsAndWeightToExerciseSet < ActiveRecord::Migration
   def self.up
-    add_column :exercise_sets, :reps, :decimal
-    add_column :exercise_sets, :weight, :decimal
+    add_column :exercise_sets, :reps, :integer
+    add_column :exercise_sets, :weight, :integer
   end
 
   def self.down
